apps/ccm-api/requirements.txt
# Core FastAPI and web dependencies
fastapi==0.115.2
uvicorn[standard]==0.35.0
sqlalchemy==2.0.42
asyncpg==0.30.0
redis==6.4.0
pydantic==2.11.7
pydantic-settings==2.10.1
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
prometheus-client==0.25.0
structlog==25.4.0
cryptography==45.0.6
gitpython==3.1.45
psutil==7.0.0
websockets==15.0.1
aiofiles==24.1.0
jinja2==3.1.6
python-dotenv==1.1.1
httpx==0.27.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
faker==33.1.0

# Core scientific computing (compatible with Python 3.13)
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
requests==2.32.4
tqdm==4.67.1

# AI Services - NVIDIA NIM only (uses httpx for API calls)
# No additional AI packages needed - using httpx already included above
